requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.31.1
pydantic==2.9.2
pandas==2.2.2
requests==2.32.3
python-dotenv==1.0.1
